+++ react/ModalContext.ts
@@ -1,0 +1,3 @@
+import { useModalDispatch } from './components/ModalContext'
+
+export default { useModalDispatch }