+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Export `useModalDispatch` hook.
+
+### Fixed
+- Modal not removing the `hiddenBody` class if it gets unmounted without beeing closed.
 
 ## [0.0.1] - 2020-01-27
 ### Added